Solution Patch, Site 추가, bug 수정

+++ teamKMS/src/main/java/com/devworker/kms/repo/solution/SolutionBugRepo.java
@@ -9,5 +9,5 @@ import org.springframework.data.repository.query.Param;
 
 public interface SolutionBugRepo extends PagingAndSortingRepository<SolutionBugDao, Long> {
 	@Query ("select t from SolutionBugDao t where t.menuId = :menuId")	
-    Page<SolutionBugDao> getBug(@Param("menuId")int menuId, Pageable pageable);
+    Page<SolutionBugDao> getBug(@Param("menuId")long menuId, Pageable pageable);
 }
